Add generic multie launcher
